+++ src/ppo/train.py
@@ -114,26 +114,33 @@ if __name__ == "__main__":
         action="store_true",
         help="render the environment on evaluation",
     )
+    train_config.add_argument(
+        "-DB",
+        "--debug",
+        action="store_true",
+        help="debug mode",
+    )
 
     args = parser.parse_args()
 
     print(colored("Initializing data folders", "blue"))
     # Init model checkpoint folder and uncertainties folder
-    if not os.path.exists("param"):
-        os.makedirs("param")
-    if not os.path.exists("uncertainties"):
-        os.makedirs("uncertainties")
-    if not os.path.exists("render"):
-        os.makedirs("render")
-    if not os.path.exists(f"render/{args.model}"):
-        os.makedirs(f"render/{args.model}")
-    else:
-        files = glob.glob(f"render/{args.model}/*")
-        for f in files:
-            os.remove(f)
-    if not os.path.exists("uncertainties/train"):
-        os.makedirs("uncertainties/train")
-    init_uncert_file(file=f"uncertainties/train/{args.model}.txt")
+    if not args.debug:
+        if not os.path.exists("param"):
+            os.makedirs("param")
+        if not os.path.exists("uncertainties"):
+            os.makedirs("uncertainties")
+        if not os.path.exists("render"):
+            os.makedirs("render")
+        if not os.path.exists(f"render/{args.model}"):
+            os.makedirs(f"render/{args.model}")
+        else:
+            files = glob.glob(f"render/{args.model}/*")
+            for f in files:
+                os.remove(f)
+        if not os.path.exists("uncertainties/train"):
+            os.makedirs("uncertainties/train")
+        init_uncert_file(file=f"uncertainties/train/{args.model}.txt")
     print(colored("Data folders created successfully", "green"))
 
     # Virtual display
@@ -228,6 +235,7 @@ if __name__ == "__main__":
         eval_interval=args.val_interval,
         model_name=args.model,
         skip_zoom=args.skip_zoom,
+        debug=args.debug
     )
 
     trainer.run()
